fix: serialize TrashType as int in route history; preserve live-tracker state on snapshot refresh
- RouteRunSummaryDto/RouteRunDetailDto: change TrashType from TrashType
  enum to int so JsonStringEnumConverter no longer emits "Mixed"/"Paper"
  etc. — Angular trashTypeLabel(t: number) was always falling through to
  'Неизв.' because string !== number comparisons always fail.
- RouteRunService.ToSummary / GetByIdAsync: cast (int)r.TrashType.
- TrucksController.GetActiveRuns anonymous projection: cast (int)r.TrashType
  so the /api/trucks/route/active snapshot also sends numbers.
- live-driver-tracking.service.ts applySnapshot: when lastPosition is null
  but the driver is in our local SignalR state, keep that live entry instead
  of silently dropping it. A tab-visibility refresh was wiping truck markers
  that SignalR had already placed on the map.

+++ BinMaps.API/Controllers/TrucksController.cs
@@ -272,7 +272,7 @@ public sealed class TrucksController : ControllerBase
                     driverId          = r.DriverId,
                     driverName        = r.DriverName,
                     areaId            = r.AreaId,
-                    trashType         = r.TrashType,
+                    trashType         = (int)r.TrashType,
                     startedAt         = r.StartedAt,
                     plannedDistanceKm = r.PlannedDistanceKm,
                     plannedMinutes    = r.PlannedMinutes,
